fileupload/requirements.txt
asgiref==3.7.2
boto3==1.34.63
botocore==1.34.63
cffi==1.16.0
cryptography==42.0.5
Django==5.0.3
djangorestframework==3.14.0
jmespath==1.0.1
pycparser==2.21
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
